perf(blockchain-link): reuse the cached chain id across evm-rpc handlers
getChainId only happened to live in the multicall util; move it, together
with the per-client cache helper, into its own module. The connect probe now
primes that cache, so getInfo and the staking pool read answer from it
instead of asking the RPC for eth_chainId again.
