+++ packages/blockchain-link/src/workers/evm-rpc/handlers/getInfo.ts
@@ -2,15 +2,13 @@ import { RESPONSES } from '@trezor/blockchain-link-types';
 import type { MessageTypes, ResponseTypes as Responses } from '@trezor/blockchain-link-types';
 
 import type { Request } from '../types';
+import { getChainId } from '../utils/client';
 
 export const getInfo = async (
     request: Request<MessageTypes.GetInfo>,
 ): Promise<Responses.GetInfo> => {
     const client = await request.connect();
-    const [blockNumber, chainId] = await Promise.all([
-        client.getBlockNumber(),
-        client.getChainId(),
-    ]);
+    const [blockNumber, chainId] = await Promise.all([client.getBlockNumber(), getChainId(client)]);
 
     const block = await client.getBlock({ blockNumber, includeTransactions: false });
